Refactored spf
